Struct args::args::OrderValidation

pub struct OrderValidation<O: Clone + Display + PartialOrd> {
    // some fields omitted
}

An implementation of the Validation trait which tests whether or not a value adheres to the provided order and bound. It should be noted that the type of bound, O, must implement Clone, Display and PartialOrd.

Example

use args::{Order,OrderValidation};

let validation = OrderValidation::new(Order::GreaterThan, 0u32);
validation.is_valid(1u32) // true
validation.is_valid(0u32) // false

if validation.is_invalid(0u32) {
    // do things
    error!("{}", validation.error(0u32));
}

Methods

impl<O: Clone + Display + PartialOrd> OrderValidation<O>

fn new(order: Order, bound: O) -> OrderValidation<O>

Creates a new OrderValidation with the provided order and bound.