Trait args::HasArgs

pub trait HasArgs: Send {
    fn args(&self) -> &mut Args;

    fn full_usage(&self, brief: &str) -> String { ... }
    fn parse<C: IntoIterator>(&mut self, raw_args: C) -> &mut Args where C::Item: AsRef<OsStr> { ... }
    fn parse_from_cli(&mut self) -> &mut Args { ... }
    fn short_usage(&self) -> String { ... }
    fn validated_value_of<T: FromStr>(&self, opt_name: &str, validations: &[Box<Validation<T=T>>]) -> Result<T, ArgsError> { ... }
    fn value_of<T: FromStr>(&self, opt_name: &str) -> Result<T, ArgsError> { ... }
}

A trait designed to bind arguments and convenience methods to a struct.

Required Methods

fn args(&self) -> &mut Args

Returns a mutable references to an Args struct.

Provided Methods

fn full_usage(&self, brief: &str) -> String

Acts as a convenience method for calling the Args implementation.

fn parse<C: IntoIterator>(&mut self, raw_args: C) -> &mut Args where C::Item: AsRef<OsStr>

Acts as a convenience method for calling the Args implementation.

fn parse_from_cli(&mut self) -> &mut Args

Acts as a convenience method for calling the Args implementation.

fn short_usage(&self) -> String

Acts as a convenience method for calling the Args implementation.

fn validated_value_of<T: FromStr>(&self, opt_name: &str, validations: &[Box<Validation<T=T>>]) -> Result<T, ArgsError>

Acts as a convenience method for calling the Args implementation.

fn value_of<T: FromStr>(&self, opt_name: &str) -> Result<T, ArgsError>

Acts as a convenience method for calling the Args implementation.