Trait args::traits::HasArgs

pub trait HasArgs: Send {
    fn args() -> Args where Self: Sized;

    fn full_usage() -> String where Self: Sized { ... }
    fn parse<C: IntoIterator>(raw_args: C) -> Result<(), ArgsError> where C::Item: AsRef<OsStr>, Self: Sized { ... }
    fn parse_from_cli() -> Result<(), ArgsError> where Self: Sized { ... }
    fn short_usage() -> String where Self: Sized { ... }
    fn usage() -> String where Self: Sized { ... }
}

A trait designed to bind associated convenience methods to a struct.

Required Methods

fn args() -> Args where Self: Sized

Returns a mutable references to an associated Args struct.

Provided Methods

fn full_usage() -> String where Self: Sized

Generates a combination of the short and verbose usage messages.

fn parse<C: IntoIterator>(raw_args: C) -> Result<(), ArgsError> where C::Item: AsRef<OsStr>, Self: Sized

Acts as a convenience method for calling the associated Args implementation.

fn parse_from_cli() -> Result<(), ArgsError> where Self: Sized

Acts as a convenience method for calling the associated Args implementation.

fn short_usage() -> String where Self: Sized

Acts as a convenience method for calling the associated Args implementation.

fn usage() -> String where Self: Sized

Acts as a convenience method for calling the associated Args implementation.